Eliminate the ugly 'using static' imports for the enum & internal class that were in DataCommandList
